pyproject.toml
[project]
name = "extract-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi~=0.115.0",
    "promplate[openai]~=0.3.4.9",
    "promptools[validation]~=0.1.3.5",
    "python-dotenv~=1.0.1",
    "ruff~=0.6.7",
    "uvicorn~=0.30.6",
]

[tool.ruff]
preview = true
line-length = 200

[tool.ruff.lint]
extend-select = [
    "F",    # flake8
    "I",    # isort
    "N",    # pep8-naming
    "W",    # pycodestyle
    "UP",   # pyupgrade
    "TRY",  # tryceratops
    "RUF",  # ruff
    "FURB", # refurb
    "C4",   # flake8-comprehensions
    "ARG",  # flake8-unused-arguments
    "PIE",  # flake8-pie
    "PTH",  # flake8-use-pathlib
    "RSE",  # flake8-raise
    "SIM",  # flake8-simplify
    "SLF",  # flake8-self
]
